backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
alembic==1.13.1
python-dotenv==1.0.1
google-generativeai==0.5.2
slowapi==0.1.9
pydantic==2.6.4
pydantic-settings==2.2.1
httpx==0.27.0
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
reportlab==4.2.5
python-multipart>=0.0.9
chromadb==0.4.24
pypdf==4.1.0
langchain-text-splitters==0.0.1
